add Pinned Comment feature (#6)
New comments get retrieved periodically, via go-reddit which uses the Reddit API, and sent through a channel.

This channel is then filled with new comments, which get pulled. For each comment we parse the flair to determine:

1. if the user is verified
2. the optional org tag the user uses
3. the user's role (can be Pro Player, or a job title on their org, etc...)
5. the user's name/username

If the user is verified, we save the comment in database - with the Post ID - to be able to retrieve it for potential later editing (will be explained).

Once we have saved the comment, we then loop over all the user types: `Riot Games` for riot employees, `proplayer` for pros, `orgstaff` for org employees and `other` which can include broadcasters, journalists and so on.

In this loop, we retrieve all comments that were saved, like previously, for this thread thanks to the Post ID. We then use those comments to format all this information, including the contents of the comments, into a string that will be the content of the pinned comment.

If there is no existing pinned comment - by checking in DB - , we create one with the content generated previously and submit it to reddit. We then save the comment ID into database to be able to make the earlier check for future comments.
If there *is* an existing pinned comment, we retrieve its ID and edit it through the reddit API with the new generated content.

---

The spreadsheet is used to transform org tags into their actual name like `100T` into `100 Thieves` and so on.